SmartAnts/backend/requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
beautifulsoup4==4.14.3
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
DateTime==6.0
Deprecated==1.3.1
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.2.9
django-allauth==65.13.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
finance-datareader==0.9.100
fonttools==4.61.1
frozendict==2.4.7
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.11
joblib==1.5.3
kiwisolver==1.4.9
lxml==6.0.2
matplotlib==3.10.8
multipledispatch==1.0.0
multitasking==0.0.12
narwhals==2.14.0
numpy==2.4.0
packaging==25.0
pandas==2.3.3
peewee==3.18.3
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pykrx==1.0.51
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
requests-file==3.0.1
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
soupsieve==2.8.1
sqlparse==0.5.5
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
websockets==15.0.1
whitenoise==6.11.0
wrapt==2.0.1
xlrd==2.0.2
yfinance==1.0
zope.interface==8.1.1
